This is a lot of API refactors for #95 as well as performance improvements

Key parts
- `contains` operation improved, relying on central equality code paths
- More error reporting improvements
  - Tone of errors
  - More context provided
- Perf
  - Avoid more copies
  - `if existence` doesn't take on expensive error reporting costs
- Correctness: `Path` / `Variable` behave as a `Vec1` now, ensuring there is at least one element

Controlling for the benchmarks being correct:
```
master:

test bench_parse_template  ... bench:      26,748 ns/iter (+/- 2,568)
test bench_parse_text      ... bench:         654 ns/iter (+/- 59)
test bench_parse_variable  ... bench:       2,505 ns/iter (+/- 1,076)
test bench_render_template ... bench:      18,286 ns/iter (+/- 1,671)
test bench_render_text     ... bench:         288 ns/iter (+/- 23)
test bench_render_variable ... bench:         856 ns/iter (+/- 74)

this branch:

test bench_parse_template  ... bench:      26,208 ns/iter (+/- 2,875)
test bench_parse_text      ... bench:         649 ns/iter (+/- 42)
test bench_parse_variable  ... bench:       2,604 ns/iter (+/- 222)
test bench_render_template ... bench:      15,983 ns/iter (+/- 1,512)
test bench_render_text     ... bench:         282 ns/iter (+/- 21)
test bench_render_variable ... bench:         628 ns/iter (+/- 309)
```

BREAKING CHANGES
- `Expression` / `Variable` now return references when evaluated
- `liquid_value::Path`s API is overhauled
- `Globals` API changed
- `contains` operation now has a bug fix which might impact behavior (item equality behavior changed)