@epage 2018-11-19
@emoon: some more performance improvements.