@epage 2018-11-19
I know you've said your app is render-heavy but we do have some parsing improvement in the works that I'm expecting will see a major improvement in there as well.